==> notebooks/1_cleaning_500_albums.ipynb <==
import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  # me muestre todas las columnas
#pd.set_option('display.max_rows', None)  # me muestre todas las filas

data = pd.read_csv('../data/raw/DB_Kaggle_top500albums.csv', encoding='latin1')

data.head()

data['Album'] = data['Album'].apply(lambda x: x.strip() if isinstance(x, str) else x)

data.columns

columns = ['toprankingorder', 'year', 'album', 'artist', 'genre', 'subgenre']

data.columns = columns

data.head()

data.info()

data = data.iloc[:, [0, 2, 3,1]]

data.head()

data.to_csv('../data/clean/DB_Kaggle_top500albums_clean.csv', index=False)
