==> notebooks/2_selenium_50horriblealbums.ipynb <==
## Import libraries

from selenium import webdriver
import pandas as pd
from selenium.webdriver.common.by import By
import time 
from selenium.webdriver.chrome.options import Options
from webdriver_manager.chrome import ChromeDriverManager

#driver configuration
opciones=Options()

opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
opciones.add_experimental_option('useAutomationExtension', False)
opciones.headless=False    # si True, no aperece la ventana (headless=no visible)
opciones.add_argument('--start-maximized')         # comienza maximizado
#opciones.add_argument('user-data-dir=selenium')    # mantiene las cookies
#opciones.add_extension('driver_folder/adblock.crx')       # adblocker
opciones.add_argument('--incognito')

PATH = 'driver/chromedriver'

driver = webdriver.Chrome()

driver.get('https://www.google.es')

## Get 50 horrible albums by brilliant artists

url = 'https://www.rollingstone.com/music/music-lists/horrible-albums-by-brilliant-artists-1234672895/'

driver.get(url)

#cookies 

rejectcookies = driver.find_element(By.XPATH,'//*[@id="onetrust-reject-all-handler"]')
rejectcookies.click()

#ranking number 

rankingnumber = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__number')
len(rankingnumber)

rankingnumber = [item.text for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__number')]
rankingnumber[0:5]

#band name  

bandname = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__title')
len(bandname)

bandname = [item.text for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__title')]
bandname[0:5]

#albumname  

albumname = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__additional--description')
len(albumname)

albumname = [item.text.replace("'","") for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__additional--description')]
albumname[0:5]

#releaseyear 

releaseyear = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__subtitle-1')
len(releaseyear)

releaseyear = [item.text for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__subtitle-1')]
releaseyear[0:5]

#rationale

rationale = driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__description')
len(rationale)

rationale = [item.text for item in driver.find_elements(By.CLASS_NAME,'c-gallery-vertical-album__description')]
rationale[0]

data = {'horriblerankingorder': rankingnumber, 'artist': bandname, 'album': albumname, 'year': releaseyear, 'rationale': rationale}

finaltable = pd.DataFrame(data)
finaltable.head()

finaltable.info()

finaltable['horriblerankingorder'] = finaltable['horriblerankingorder'].astype('int64')

finaltable['horriblerankingorder'].dtype

finaltable.shape

#since the information on band and release_year will come from Spotify, I will select the columns I am interested in:

finaltable = finaltable.iloc[:, [0, 2, 1, 4, 3]] 

finaltable.head()

finaltable.info()

finaltable.to_csv('../data/clean/WS_50horriblealbums_clean.csv', index=False)
