==> notebooks/4_api_albums.ipynb <==
### Importing libraries

import pandas as pd
import numpy as np
import time 

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  # me muestre todas las columnas
#pd.set_option('display.max_rows', None)  # me muestre todas las filas

### Getting target list 

all_albums = pd.read_csv('../data/clean/all_550_albums_clean.csv', encoding='latin1')

all_albums.head(3)

#### list of albums

albums = all_albums['album'].unique().tolist()

#some had the same name (e.g. Greatest Hits, Let It Be)

len(albums)

#### list of albums_artists 

album_artist = all_albums['album_artist'].unique().tolist()

len(album_artist)

### Spotify Connection 

#%pip install pyarrow
#%pip install joblib
#%pip install tqdm
#%pip install spotipy
#%pip install python-dotenv
#%pip install requests 

import requests as rq
import pyarrow
from joblib import Parallel, delayed
from tqdm import tqdm
import pymongo
from passwords import *

# Obtener el TOKEN

AUTH_URL = "https://accounts.spotify.com/api/token"

creds = {
    "grant_type": "client_credentials",
    "client_id": Client_ID,
    "client_secret": Client_secret
}

response = rq.post(AUTH_URL, 
                         data=creds, 
                         headers={"Content-Type": "application/x-www-form-urlencoded"})


TOKEN = response.json()['access_token']

# Conexión con la API

url = "https://api.spotify.com/"

headers = {"Authorization": f'Bearer {TOKEN}'}

response = rq.get(url,headers=headers)
response

### Getting Album Information 

albums[0:3]

#understanding the json 

url = 'https://api.spotify.com/v1/search?q=album%3Apet+sounds&type=album&limit=50&offset=0'

res = rq.get(url, headers=headers)
res.status_code 

data = res.json()

type(data)

data.keys()

data['albums'].keys()

data['albums']['items'][0].keys()

#quiero el id del album en Spotify 
data['albums']['items'][0]['id']

#quiero el nombre del album en Spotify
data['albums']['items'][0]['name']

#quiero el nombre del artista del album 
data['albums']['items'][0]['artists'][0]['name']

#quiero el id del artista del album 
data['albums']['items'][0]['artists'][0]['id']

#quiero el release_date del album 
data['albums']['items'][0]['release_date']

#quiero el total de tracks del album 
data['albums']['items'][0]['total_tracks']

# Replace spaces with the '+' character in the album names
insertquery = [x.replace(' ', '+') for x in albums]

# Construct a list of URLs for querying album information
album_urls = [f'https://api.spotify.com/v1/search?q=album%3A{album}&type=album&limit=50&offset=0' for album in insertquery]

# Create a list of dictionaries with album information and corresponding URLs
album_url_list = [{'album': album, 'url': url} for album, url in zip(albums, album_urls)]

album_url_list[1]

# List to store the retrieved album data
album_data_list = []

# Function to search and retrieve album information
def search_spotify_info(album_url):
    time.sleep(0.30)  # Pause between queries to avoid getting banned

    try:
        response = rq.get(album_url['url'], headers=headers)
        response.raise_for_status()  # Check for request errors
        data = response.json()
        rjson = data.get('albums', {}).get('items', [])

        for index, item in enumerate(rjson):
            if item['name'].lower() == album_url['album'].lower():
                album_data_list.append({
                    'albumnamers': album_url['album'],
                    'albumidspotify': item['id'],  # Spotify ID for the album
                    'albumnamespotify': item['name'],
                    'artistnamespotify': item['artists'][0]['name'],
                    'artistidspotify': item['artists'][0]['id'],
                    'releasedate': item['release_date'],
                    'totaltracks': item['total_tracks']
                })
                
                return rjson[index]

        # If the album is not found, raise an exception to return None
        raise Exception("Album not found")
    
    except Exception as e:
        print(f"An error occurred: {e}")
        album_data_list.append({
                    'albumnamers': None, 
                    'albumidspotify': None, 
                    'albumnamespotify': None, 
                    'artistnamespotify': None,
                    'artistidspotify': None,
                    'releasedate': None,
                    'totaltracks': None 
        })
        return {
            'album_type': None,
            'artists': None,
            'available_markets': None,
            'external_urls': None,
            'href': None,
            'id': None,
            'images': None,
            'name': None,
            'release_date': None,
            'release_date_precision': None,
            'total_tracks': None,
            'type': None,
            'uri': None
        }
    
# Execute the function for each album URL
for album_url in album_url_list:
    search_spotify_info(album_url)

# Create a DataFrame from the list of dictionaries
dataalbums = pd.DataFrame(album_data_list)

dataalbums.dropna(inplace=True)

dataalbums.shape

dataalbums.head()

dataalbums.to_csv('../data/clean/api_albums_clean.csv', index=False)
