==> notebooks/10_sql_table_albums.ipynb <==
## Import libraries

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  # me muestre todas las columnas
#pd.set_option('display.max_rows', None)  # me muestre todas las filas

## Import files

sqltable_albums = pd.read_csv('../data/clean/api_albums_clean.csv', encoding='latin1')
sqltable_albums.head(2)

sqltable_albums.shape

## Clean NAs and duplicates

#clean NAs (some albums available in the RS DB were not found in Spotify)
sqltable_albums.dropna(inplace=True)

sqltable_albums.shape

sqltable_albums.duplicated().any()

## Column names

sqltable_albums.columns

finalcolumns = ['rs_albumname', 'sp_albumid', 'sp_albumname', 'sp_artistname', 'artists_sp_artistid','sp_releasedate','sp_totaltracks']

sqltable_albums.columns = finalcolumns

sqltable_albums.head(1)

## Reoder columns

sqltable_albums = sqltable_albums.iloc[:, [1, 2, 5, 6, 4]]
sqltable_albums.head(1)

## Data type

sqltable_albums.info()

#sqltable_albums['sp_releasedate'] = pd.to_datetime(sqltable_albums['sp_releasedate'], format='%Y')

sqltable_albums['sp_releasedate'].dtype

sqltable_albums['sp_totaltracks'] = sqltable_albums['sp_totaltracks'].astype(int)

sqltable_albums['sp_totaltracks'].dtype

## Export 

import mysql.connector as conn
from sqlalchemy import create_engine

with open('../pw.txt') as file: 
    
    password = file.read()

str_conn=f'mysql+pymysql://root:{password}@localhost:3306/rsalbums'

cursor = create_engine(str_conn)

sqltable_albums.to_sql(name='albums',
              con=cursor,
              if_exists = 'replace',
              index=False)

sqltable_albums.to_csv('../data/sqlexports/sqltable_albums.csv', index=False, encoding='latin1')
