==> notebooks/11_sql_table_artists.ipynb <==
## Import libraries

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  # me muestre todas las columnas
#pd.set_option('display.max_rows', None)  # me muestre todas las filas

## Import files

sqltable_artists = pd.read_csv('../data/clean/api_artists_clean.csv', encoding='latin1')
sqltable_artists.head(2)

sqltable_artists.shape

## Clean NAs and duplicates

sqltable_artists.dropna(inplace=True)

sqltable_artists.shape

sqltable_artists.duplicated().any()

import re

for col in sqltable_artists.columns:
    sqltable_artists[col] = sqltable_artists[col].apply(lambda x: re.sub(r'[^\w\s]', '', x) if isinstance(x, str) else x)

sqltable_artists.head()

## Column names

sqltable_artists.columns

finalcolumns = ['sp_artistid', 'sp_artistname', 'sp_followers', 'sp_popularity', 'sp_genres']

sqltable_artists.columns = finalcolumns

sqltable_artists.head(1)

## Reoder columns

sqltable_artists = sqltable_artists.iloc[:, [0, 1, 2, 3]]
sqltable_artists.head(1)

## Data type

sqltable_artists.info()

## Export 

import mysql.connector as conn
from sqlalchemy import create_engine

sqltable_artists.head(2)

with open('../pw.txt') as file: 
    
    password = file.read()

str_conn=f'mysql+pymysql://root:{password}@localhost:3306/rsalbums'

cursor = create_engine(str_conn)

#sqltable_artists.to_sql(name='artists',
              #con=cursor,
              #if_exists = 'append',
              #index=False)

sqltable_artists.to_csv('../data/sqlexports/sqltable_artist.csv', index=False, encoding='latin1')
