==> notebooks/12_sql_table_tracks.ipynb <==
## Import libraries

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)  # me muestre todas las columnas
#pd.set_option('display.max_rows', None)  # me muestre todas las filas

## Import files

sqltable_tracks = pd.read_csv('../data/clean/api_tracks1_clean.csv', encoding='latin1')
sqltable_tracks.head(2)

sqltable_tracks.shape

## Clean NAs and duplicates

sqltable_tracks.dropna(inplace=True)

sqltable_tracks.shape

sqltable_tracks.duplicated().any()

## Column names

sqltable_tracks.columns

finalcolumns = ['sp_trackid', 'sp_trackname', 'albums_sp_albumid', 'artists_sp_artistid', 'sp_artistname']

sqltable_tracks.columns = finalcolumns

sqltable_tracks.head(1)

## Reoder columns

sqltable_tracks = sqltable_tracks.iloc[:, [0, 1, 2]]
sqltable_tracks.head(1)

## Data type

sqltable_tracks.info()

## Export 

import mysql.connector as conn
from sqlalchemy import create_engine

sqltable_tracks.head(1)

with open('../pw.txt') as file: 
    
    password = file.read()

str_conn=f'mysql+pymysql://root:{password}@localhost:3306/rsalbums'

cursor = create_engine(str_conn)

sqltable_tracks.to_sql(name='tracks',
              con=cursor,
              if_exists = 'replace',
              index=False)

sqltable_tracks.to_csv('../data/sqlexports/sqltable_tracks.csv', index=False, encoding='latin1')
